requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
arrow==0.14.7
    # via ics
asgiref==3.5.2
    # via django
beautifulsoup4==4.11.1
    # via bvg-grabber
bvg-grabber==0.2.1
    # via -r requirements.in
certifi==2022.5.18.1
    # via requests
charset-normalizer==2.0.12
    # via requests
django==4.0.4
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-bootstrap3
    #   django-json-rpc
    #   django-node
    #   django-webpack
    #   djangoajax
django-appconf==1.0.5
    # via -r requirements.in
django-bootstrap3==21.2
    # via -r requirements.in
django-json-rpc==0.7.2
    # via -r requirements.in
django-node==4.0.0
    # via django-webpack
django-tagging==0.5.0
    # via -r requirements.in
django-webpack==3.1.1
    # via -r requirements.in
django-webpack-loader==1.5.0
    # via -r requirements.in
djangoajax==3.3
    # via -r requirements.in
feedparser==6.0.10
    # via -r requirements.in
gcm==0.3
    # via -r requirements.in
htmlparser==0.0.2
    # via -r requirements.in
ics==0.7
    # via -r requirements.in
idna==3.3
    # via requests
optional-django==0.1.0
    # via django-webpack
paho-mqtt==1.6.1
    # via -r requirements.in
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   python-ldap
pyasn1-modules==0.2.8
    # via python-ldap
python-dateutil==2.8.2
    # via
    #   arrow
    #   bvg-grabber
    #   ics
python-ldap==3.4.0
    # via -r requirements.in
python-mpd2==3.0.5
    # via -r requirements.in
requests==2.27.1
    # via
    #   bvg-grabber
    #   django-node
sgmllib3k==1.0.0
    # via feedparser
six==1.16.0
    # via
    #   bvg-grabber
    #   django-json-rpc
    #   ics
    #   python-dateutil
soupsieve==2.3.2.post1
    # via beautifulsoup4
sqlparse==0.4.2
    # via django
tatsu==5.6.1
    # via ics
urllib3==1.26.9
    # via requests
